span_ext mod
